Move elwing libs to main js libs folder.
