+++ javascriptv3/example_code/libs/utils/util-string.js
@@ -2,20 +2,39 @@
  * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
  * SPDX-License-Identifier: Apache-2.0
  */
+import { pipe, adjust, join, split, toLower, map, toUpper } from "ramda";
+
 import { v4 as uuidv4 } from "uuid";
 
-export const getUniqueName = (name) => `${uuidv4()}-${name.toLowerCase()}`;
+const getUniqueName = (name) => `${uuidv4()}-${name.toLowerCase()}`;
 
-/**
- *
- * @param { string } source the string to modify
- * @param { string } str the string to affix to the source
- * @returns { string } the postfixed source
- */
-export const postfix = (source, str) => {
+const postfix = (source, str) => {
   if (typeof str !== "string") {
     throw new Error("Cannot postfix a non-string value.");
   }
 
   return `${source}${str}`;
+};
+
+const downcaseSplit = pipe(toLower, split("-"));
+
+const capitalize = pipe(Array.from, adjust(0, toUpper), join(""));
+
+const kebabCase = pipe(downcaseSplit, join("-"));
+
+const pascalCase = pipe(downcaseSplit, map(capitalize), join(""));
+
+const snakeCase = pipe(downcaseSplit, join("_"));
+
+const titleCase = pipe(downcaseSplit, map(capitalize), join(" "));
+
+export {
+  capitalize,
+  downcaseSplit,
+  getUniqueName,
+  kebabCase,
+  pascalCase,
+  postfix,
+  snakeCase,
+  titleCase,
 };
